requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
email-validator==2.2.0
python-dotenv==1.0.1
click==8.1.7
python-dateutil==2.8.2
openpyxl==3.1.2
fpdf==1.7.2
blinker==1.8.2
colorama==0.4.6
dnspython==2.6.1
alembic==1.13.3
psycopg2-binary==2.9.9
Pillow==10.1.0
gunicorn==20.1.0
redis==5.0.1